fix variable type

+++ src/constructor/keyboard_handlers/delete_route_keyboard.py
@@ -1,6 +1,6 @@
 
 from src.database.routes import delete_user_route, get_route_by_id_and_username
-from src.constructor.services.tg_keyboard import handle_navigation_buttons
+from src.constructor.services.tg_keyboard import handle_navigation_buttons, handle_route_info_button
 
 
 def handler(keyboard_id, callback_query, chat_state):
@@ -12,4 +12,10 @@ def handler(keyboard_id, callback_query, chat_state):
     username = callback_query["from"]["username"]
     route_id = callback_query['data']
     route = get_route_by_id_and_username(route_id, username)
+    chat_id = callback_query["from"]["id"]
+
+    if button_info.startswith("info"):
+        handle_route_info_button(route, chat_id)
+        return
+
     delete_user_route(route["pk"], route["sk"])
